+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Refactor SceneManager into static class (it was a ScriptableObject)
+
 ### Removed
 - ISceneManager interface
 - Scriptable Settings Provider package dependency